support the Content-Length header for reading files
We now send the Content-Length header.
In case of a complete read, it's the filesize,
in case of a partial read, it's the concatenated length
of all parts that were requested (and that we're trying to push)

Also we now send a 206 status code when only parts of the file are
returned. (When parts were requested, but the whole file returned,
it's 200).

We don't have the content-range yet, because I'm not sure what to
do with multipart requests there.