add StorageDir and StorageFile classes; update dir listing
the storage API has to send StorageDir and StorageFile
objects back, instead of the plain object names.
This allows to acommodate more information in the results and
gives a standard class to use.

The http rendering is updated to show the extra information
and links to follow through the filesystem.

The get_file function is updated to redirect if the client
tries to download a directory instead of a file.