get user metadata information from irods -> storage.stat()
implement the storage.stat() call which get metadata from the storage.
in the irods case, this are some attributes we have to get
specifically from the file or dir handle, and the user metadata.

they are integrated in the json output of a dir listing.

to allow this, the StorageDir and StorageFile classes
had to be changed.