use the metadata api for gettting metadata
metadata management should be distinct from the storage part
to allow other components, which are able to store richer metadata
than the storage solution, to be used.

The metadata.py should also be able to combine metadata from different
sources.

This first implementation takes metadata from the only available source,
the storage, which is, at the moment, irods.