big commit: add support for CDMI URL arguments
Several changes:
1. parse cdmi command line arguments from the URI
    none have values, except metadata and children
    (md fields and range, respectively)
2. the cdmi json generator returns functions instead of
    values so that arguments from the URI can be added
    (e.g. childrenrange)
3. filter the cdmi json generator result by the arguments
    from the URI
    print all if no arguments, filter otherwise

This way is very much inspired by how stoxy does it:
https://github.com/stoxy/stoxy/blob/master/stoxy/server/endpoint/cdmi/view.py

What this additionally supports is sending back a specified
childrenrange. For this we had to implement the parital
functions, which take their last arguments from the URI
arguments.

In this version, filtering CDMI metadata does not work yet.