fix the get_user_metadata function
After the refactoring it returned a dict iwth key user_metadata,
whose value was a dict with the metadata as returned from irods.

Now it returns, as it should, a dict directly with the metadata
value in it. The irods metadata elements, which can have two values,
and converted from (key, val1, val2) to (key, (val1, val2)).
This seems the most sane approach to me.