support cdmi value range as replacing range headers
If a request wants CDMI back, it now looks at the value range
parameter, and _not_ the range header, for the range of the file
to get.
Currently, it only parses a single range, because I am not sure
if multi ranges are supposed to be supported.

Both functions get_cdmi_file/dir_obj have been modified to check the
cdmi parameters (if cdmi was requested) and fail before contacting
the storage backend (or doing anything else, for that matter).