restructure the project to use blueprints
In this commit almost all files had to be touched.

The project is now clustered into two blueprints:
- registration
- http_storage

The registration blueprint takes care of all the eudat-specific
functions, the http storage one represents the CDMI endpoint
(with multiple backends).

To separate the blueprints, the files have been put into two folders.
The shared modules are auth, metadata, and common.

The auth module will remain shared, but I can think of a third
blueprint for metadata if this becomes more complex
(invenio or so ...).

There are no direct dependencies between the two blueprints, but strong
interaction between metadata and http_storage.

Additionally, the app object is now created through a factory function,
which made changes in several files and the tests necessary, as the
'app' object is no longer available freely.
Modules inside a blueprint can call the flask.current_app object,
but this is not always available.
Including any file that depends on storage.py needs an accessible
app object.

The name for the http_storage is not the best, I'd be happy to hear
if you have a better one! It should convey that it's a storage server,
that it speaks HTTP, that it supports CDMI, but that it's not
limited to it.