workaround for the irods close_storage() function and blueprints
moving to blueprints created a problematic circular reference
from irodsstorage.py to http_storage/__init__.py.
It is needed to hook the close_storage() function into the
teardown of a request.

This workaround (which is not nice, because it spreads intermediate
functions all over the place) registers a @teardown function in
__init__.py, which calls cdmi.teardown, which calls storage.teardown.
This also prevents the function from having a more meaningful title.