add a gimmick to suppress irods output in terminals
add a wrapper around the irods.clientLogin*() function
to slience it.
This makes for nicer output of test runs and interactive runs.

This is dones by a hack to redirect stdout and stderr to devnull
when calling this function and putting them back afterwards.

Since it's only used when setting up connections, and we
cache connections, it should not affect performance much.