fix the check_path function in localstorage to accept '/'
the check with os.path.normpath would fail for '/', because
normpath does not transform it to '' (which is good).
Because it removes the trailing slash for everything else,
we had a switch that compares only path[:-1] to the normpath
result if the last char was a '/'.
Now the case is:
  remove last char if '/' and if len path > 1