fix the irods connection reuse; more debug messages
irods connections were not reused, because we forced
creating a new connection whenever the used connections
are 0. It should be only when the available connections,
the ones in the applicable user pool, are 0.

add more debug messages and display the current sizes
of a user pool and of the usedconnections set in the
debug messages.