refactor the use of irods connections
this patch changes:
1. the use of multiple connections in one request is now safe
2. irods connection do not make use of the g object,
  the request stack
3. we do not need a teardown_request function anymore

This changes the way irods connections are handled.
The authenticate function no longer puts a connection on the g stack,
but stores it in the connection pool. The rest of the API functions
take it from there, and can also initiate the creation of more
connections, if necessary.