remove forcing authentication on every API request
the localstorage api function had a with_auth decorator
that would redo the auth on every call. Change it so that
we don't do it unnecessarily and that unit testing becomes easier.

I also added a little hack (beware ...) ot the storage
interface tests not to test for wrong users if they run with the
local API. It might be turned around eventually saying:
"Only test for authentication when using the irods backend."