make the blueprint initializations explicit
python's model of implicitly importing __init__.py files
makes it impossible to prevent the blueprints to be imported
when anything else in the blueprint folder gets imported.

this makes it hard to write initialization functions, like
the database creation script db_create.py or running unit tests.

by making the initialization explicit and changing the main
__init__.py accordingly, we simplify those cases.