Fixing db in workers.
Thanks to blueprints it did not work out of the box.

Second problem are the sessions in multithreaded environment. Flask create session for each request but we need access to db from registration workers. Explicitly closing session in POST handle then. This might be a problem in the future.

Clean up idea: use queue.Queue to store request_ids and a worker pool.