conditionally only load parts of the app
New configuration parameters decide whether:
the registration component should be loaded,
the storage should be provide read functions,
or the storage should provide write and delete functions.

This enables the app to be deployed distributed, with
each machine having different responsibilites.

The split of the storage interface into read and write
makes is easy to expose a storage read-only by not allowing
any requests that can make modifications.