hackily implement CDMI PUT copy
Parse a cdmi request enough to find out
if it is a copy request, then try to fulfill it.
All other cdmi requests will do nothing.

This copy implementation does not treat the source uri
as cdmi, but accesses it in plain http (with cdmi headers).
This shortcut was taken to avoid full parsing of the
cdmi message body.

Because it may or may not contain a possibly large
value field, we will have to parse the json partially,
and I am not aware of any json library that could
do this for us.

As usual, the auth tokens are taken from the incoming
request.

+ it can copy from non-cdmi sources :)