split get_obj functions into different modules per accept-type
trial the separation of access functions into multiple modules.
I started with http_storage.init.get_(cdmi)_obj, which calls
get_(cdmi)_dir_obj and get_(cdmi)_file_obj.

The idea is simple:
Each module is responsible for one response type, now cdmi, html and
plan file content, or json. Each provides the functions get_file_obj
and get_dir_obj.
The route function checks the accept header or other criteria to
choose one and then provides a variable with the module. Then
module.get_dir|file_obj is called in the function.

cdmi and noncdmi provide both get_file_obj and get_dir_obj, so
no problems can occur, json only provides displaying a dir.
If it gets to be issued on a file, the decorator that does the accept
matching catches the AttributeError (because the function doesn't exist
in the module) and returns a 406 Not Acceptable.

For the moment, this does not support content negotiation very well,
('accept: application/json, */*;q=0.9' would fail). I am sure this can
be extended in due course and also that probably not many API would be
able to handle html, if json is not available or vice versa ...

The good part: it's easy to add new output types.
1. write the module, say xml.py with the output functions
  get_file_obj, get_dir_obj, put_file_obj, ...
2. go to http_storage/init.py and
    - import the new module
    - edit the decorator; add a clause
        elif request_wants_xml():
          access_module = xml
3. if necessary, implement the request_wants function in
  common.py

Now requests looking only for xml will be served by the new module,
for requests that could also satisfy other types it depends where you
put the elif in the decorator.
For functions you haven't implemented, the client will never see an
exception, but a 406 http error.

For a type that implements additional functions it's slightly more
difficult, but not much. All shared methods like GET, PUT, and DELETE go
like the others. For MKCOL and co you can either:
- add them as shared methods to init.
- write the routing registration in your module and add them to the
  blueprint
- add a new blueprint with the function routing and add them in
  create_app()

A similar route should be taken for the cdmi-specific URIs like
cdmi_capabilities.

An issue not addressed in this commit is how to deactivate parts of the
output, i.e. switch off all cdmi-related capabilities. This can be
implemented by directly looking at the config variables in the selecting
decorator, but maybe there is a better way.