add parameters to switch off output types
Enable the option to switch off the CDMI and json output type with
an ACTIVATE_* param in the config.

If the choose_access_type function finds those required, then it checks
in the config for this variable. If it's False or not set, it won't
link the required access type module and the client will receive
a 406.

This is the simplest possible solution and the type of variables used
can be changed, for example to a list of supported output types:
ACCESS_TYPES = ['cdmi', 'json', 'html', webdav']
and automatically activating the ACTIVATE_STORAGE parts if one is set.

The content negotiation now does not work up to RFC. The specific
type json is only used if no other type has the same (or higher)
quality. We don't look out for the positioning.
e.g. 'accept: application/json, */*' gives an html response.

CDMI does not do content-negotiation at all, it only relies on the
CDMI version header.