add static cdmi_capabilities methods
Add implementations of the cdmi_capabilities URIs for system-wide
capabilities, for object, and containers.
This first implementation is static and considers the capabilities that
we already support (= True) and the ones we consider (= False).
Everything else is not present and thus disabled. For a copmlete
reference see the cdmi spec.

This is an implementation of additional routes added for a specific
access type. To enable them, they had to be included in the general
__init__ as new Blueprint.
The same route can be taken to implement other methods, such a for
webdav.