create a config variable for the scratch space
let the scratch space be defined by a configuration variable.
in an earlier version, it would point to the root dir '/', which
is not correct. an admin should be able to define a specific location
set up as scratch space, where this link should go.

consequently, we updated the templates and the template calls to
use the SCRATCH_SPACE variable.

this is open for discussion, and this commit probably does not
include the best solution (e.g. link to another webserver for scratch,
have multiple scratch spaces, work without the config var specified, ...)