add first support for the CDMI objectID
implement objectID support for files and directories
(aka objects and containers).

A CDMI objectID will be created on upload based on:
- the CDMI_ENTERPRISE_NUMBER config variable
- randomness

It's stored on the backend with set_user_metadata, this
should work for irods, but is not supported in the
localstorage, then it'll return null as before.

An open problem is the objectID assigned to the parent of the
root element ('/' or '/anything' with partial fs access).
I could not find anything about that in the CDMI spec, so now
the parentID given for those elements is null.
(this breaks the cdmi tests)