allow any value for the objectID of the root ('/') parent
the cdmi spec does not say anything if the parent of the root dir
should have an objectID, but it suggests that all entries, called
by name, so also the root entry, should have a parent path and parentID.
in the case that the root dir is not '/', there might be even some sense
in displaying e.g. /srv/ for the root: /srv/01/ to construct the full
path from both.
this patch relaxes the tests so that we don't check the parentID of the
root dir, because there is no clear way of handling it.