fix the CDMI for larger inputs (still slow and mem-consumptive)
fix reading in the cdmi put object using the ijson library.
this is a library to stream json input, but it still processes values as
complete values, not as stream. In our case, where the 'value' can be
quite large, it (1) consumes lots of time and space with the python
backend, or (2) fails with the yajl backend, if the 'value' size is
larger than its internal buffer.

I think the right way to go is to modify the python backend to support
returning a generator instead of a plain string value, but that might
take more time to find all places that have to be changed to avoid any
complete copy in memory.