make the registrationworker access the storage through http
the registration worker no longer directly connects to irods, but goes
through this http/cdmi interface.
to register a file, it uses the CDMI PUT request with a 'copy' source
url body, as specified in the cdmi docs.
to start the replication, it adds the .replicate file to the storage,
again using the CDMI PUT request, but with a 'value' body.
this still needs to be refined if the replication should also work for
other backends that have other mechanisms to start a replication.
(e.g. through an endpoint /replicate in the storage interface)

along with these changes, the cdmiclient is updated to support the
cdmi_copy command doing CDMI PUT with 'copy' and to support a streaming
GET request.
(which will be used for the streaming upload for registration)

it also updates the 'destination' used to be valid HTTP URL and adds a
config parameter for that ("HTTP_ENDPOINT") that points to the (local)
http-api.
Last, the config parameters for the Epic handle system are reduced to
all follow the schema 'HANDLE_<value>'.