fix the irods url created for replication
use the urlparse module's ParseResult type to represent the destination
url in the context. this allows us to easily take parts of it and create
the irods scheme replication url. we can also use urlparse and
urlunparse to switch between string and complex representations.
it doesn't support the irods:// scheme though, so we have to construct
that by hand.