adapt the storage tests to the refactored authentication
the authentication now consistently returns the http error code 401 in
case of missing or incorrect credentials and the tests have been adapter
to that.
this is conform to the http/1.1 spec and the CDMI spec.

(before, 403 errors were returned after an unsuccessful
authentication attempt)

add the app's secret key to test config that is needed for the
flask-login plugin