optimize reading a file through cdmi
only go to the backend, when the client asks for the value.
all other attributes (including the file size that the read()
function provides, but that we don't use anymore) can be provided
by a stat. this avoid opening the file at all.
note: also before there was no unnecessary data transferred, we now
just also skip the file opening step)

this change does _not_ change the interface to the storage, but
removing the 'file_size' return value from read() would be in order.