extend the PUT 3rd-party-copy with a local-only copy
The CDMI PUT with a src-url can now copy items locally if both src and
destination are on the same storage. It tries it by passing the path
part of the src url and the destination path to the copy_file_obj()
function that performs the local copy. If this fails because the src is
not a local file, or because the copy mechanism fails for other reasons,
it falls back to the streaming copy through the http interface.

This is implemented for irods and localstorage, dmlite doesn't provide
this capability.

current drawbacks:
- it doesn't distinguish perfectly between errors, so even if the permission
  on the (local) src was denied, it will try again with a sterming copy.
  On irods we have no other possibility thanks to the irods api, on the
  localstorage it might be better.
- There are ambiguities when a file on a remote site has the same path
  as a local one -- the local file would always be preferred.
  I don't see a security problem, though, because the user invoking the
  copy still needs read access on the local file that is 'accidentally'
  copied.
  With the information available, I couldn't think of any reliable
  mechanism to distinguish a file on the local backend and from remote
  ones, considering that a local backend can show itself with more than
  one server name and that a remote storage can still be under the local
  domain.

Improvements:
- being able to check a file's checksum for the local copy would be
  perfect: if the checksum matches, we can copy the local file,
  otherwise we take the remote one, but we have to investigate if we can
  rely on this information being available in all cases.