add a check before starting the local copy
check if the src location includes the set CDMI domain before attempting
a local (irods-internal) copy. this helps avoiding a conflict if one
wants to copy from a remote source, while the same path exists locally.
The check expects that the remote source location (that is:
server.domain.tld) does not include the string that is set as CDMI
domain.